HDDS-12995. Extract --prefix from ListOptions
