+++ hadoop-ozone/cli-admin/src/main/java/org/apache/hadoop/ozone/admin/nssummary/DiskUsageSubCommand.java
@@ -34,6 +34,7 @@ import org.apache.commons.io.FileUtils;
 import org.apache.hadoop.hdds.cli.HddsVersionProvider;
 import org.apache.hadoop.hdds.server.JsonUtils;
 import org.apache.hadoop.ozone.shell.ListOptions;
+import org.apache.hadoop.ozone.shell.PrefixFilterOption;
 import picocli.CommandLine;
 
 /**
@@ -68,6 +69,9 @@ public class DiskUsageSubCommand implements Callable {
 
   @CommandLine.Mixin
   private ListOptions listOptions;
+
+  @CommandLine.Mixin
+  private PrefixFilterOption prefixFilter;
 
   private static final String ENDPOINT = "/api/v1/namespace/du";
 
@@ -147,7 +151,7 @@ public class DiskUsageSubCommand implements Callable {
         printWithUnderline("DU", true);
         printDUHeader();
         int limit = listOptions.getLimit();
-        String seekStr = listOptions.getPrefix();
+        String seekStr = prefixFilter.getPrefix();
         if (seekStr == null) {
           seekStr = "";
         }
