+++ hadoop-ozone/cli-shell/src/main/java/org/apache/hadoop/ozone/shell/bucket/ListBucketHandler.java
@@ -27,6 +27,7 @@ import org.apache.hadoop.ozone.client.OzoneClient;
 import org.apache.hadoop.ozone.client.OzoneVolume;
 import org.apache.hadoop.ozone.shell.ListOptions;
 import org.apache.hadoop.ozone.shell.OzoneAddress;
+import org.apache.hadoop.ozone.shell.PrefixFilterOption;
 import org.apache.hadoop.ozone.shell.volume.VolumeHandler;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
@@ -42,6 +43,9 @@ public class ListBucketHandler extends VolumeHandler {
   @CommandLine.Mixin
   private ListOptions listOptions;
 
+  @CommandLine.Mixin
+  private PrefixFilterOption prefixFilter;
+
   @CommandLine.Option(names = {"--has-snapshot"},
       description = "Only show buckets that have at least one active snapshot.")
   private boolean filterByHasSnapshot;
@@ -54,7 +58,7 @@ public class ListBucketHandler extends VolumeHandler {
     ObjectStore objectStore = client.getObjectStore();
     OzoneVolume vol = objectStore.getVolume(volumeName);
     Iterator<? extends OzoneBucket> bucketIterator =
-        vol.listBuckets(listOptions.getPrefix(),
+        vol.listBuckets(prefixFilter.getPrefix(),
             listOptions.getStartItem(), filterByHasSnapshot);
     List<Object> bucketList = new ArrayList<>();
     int counter = 0;
