+++ hadoop-ozone/cli-shell/src/main/java/org/apache/hadoop/ozone/shell/volume/ListVolumeHandler.java
@@ -25,6 +25,7 @@ import org.apache.hadoop.ozone.client.OzoneVolume;
 import org.apache.hadoop.ozone.shell.Handler;
 import org.apache.hadoop.ozone.shell.ListOptions;
 import org.apache.hadoop.ozone.shell.OzoneAddress;
+import org.apache.hadoop.ozone.shell.PrefixFilterOption;
 import org.apache.hadoop.ozone.shell.Shell;
 import org.apache.hadoop.security.UserGroupInformation;
 import picocli.CommandLine;
@@ -47,6 +48,9 @@ public class ListVolumeHandler extends Handler {
 
   @CommandLine.Mixin
   private ListOptions listOptions;
+
+  @CommandLine.Mixin
+  private PrefixFilterOption prefixFilter;
 
   @Option(names = {"--user", "-u"},
       description = "List accessible volumes of the user. This will be ignored"
@@ -71,10 +75,10 @@ public class ListVolumeHandler extends Handler {
     Iterator<? extends OzoneVolume> volumeIterator;
     if (userName != null && !listOptions.isAll()) {
       volumeIterator = client.getObjectStore().listVolumesByUser(userName,
-          listOptions.getPrefix(), listOptions.getStartItem());
+          prefixFilter.getPrefix(), listOptions.getStartItem());
     } else {
       volumeIterator = client.getObjectStore().listVolumes(
-          listOptions.getPrefix(), listOptions.getStartItem());
+          prefixFilter.getPrefix(), listOptions.getStartItem());
     }
 
     int counter = printAsJsonArray(volumeIterator, listOptions.getLimit());
